requirements.txt
pillow
pyautogui
pytesseract
numpy
opencv-python
keyboard